Add files via upload
This is a loan approval dataset. It is a collection of financial records and associated information used to determine the eligibility of individuals or organizations for obtaining loans from a lending institution. it contains serveral parameters such as cibil score, income, employment status, loan term, loan amount, assets value, and loan status.